pyproject.toml
[project]
name = "cci"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
  { name = "", email = "" },
  { name = "Halvor Kvamme", email = "flkz@proton.me" },
]
dependencies = [
    "numpy>=1.26.4",
    "optuna>=3.5.0",
    "torchmetrics>=1.3.1",
    "jupyter>=1.0.0",
    "polars[pyarrow]>=0.20.15",
    "scipy>=1.12.0",
    "matplotlib>=3.8.3",
    "tensorboard>=2.16.2",
    "scikit-learn>=1.4.1.post1",
    "plotly>=5.19.0",
    "pandas>=2.2.1",
    "DeepDiff>=6.7.1",
    "hiplot>=0.1.33",
    "rerun-sdk>=0.14.1",
    "typer[all]>=0.9.0",
    "structlog>=24.1.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }


[project.optional-dependencies]
visualize = ["optuna-dashboard>=0.15.0"]
[tool.ruff]
line-length = 120

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
files = "src"

[tool.pdm]
plugins = ["pdm-plugin-torch>=23.4.0"]
distribution = false

[tool.pdm.dev-dependencies]
lsp = [
  "pylsp-mypy>=0.6.8",
  "python-lsp-server>=1.9.0",
  "pylsp-rope>=0.1.11",
  "python-lsp-ruff>=2.0.1",
]
dev = ["bandit>=1.7.6", "mypy>=1.8.0", "ruff>=0.1.11", "nbstripout>=0.7.1"]

[tool.pdm.scripts]
bandit_ = "bandit --r -ll src"
mypy_ = "mypy src/"
lint.composite = ["mypy_", "bandit_"]
fmt = "ruff format src"

cci = "python cci"

[tool.pdm.plugin.torch]
dependencies = ["torch==2.2.1", "torchvision==0.17.1"]
lockfile = "torch.lock"
enable-cpu = true

enable-rocm = true
rocm-versions = ["5.7"]

enable-cuda = true
cuda-versions = ["cu118"]
